Adding Mods folder for feature examples, factoring physics rag doll 2D HUD into a mod
