+++ FeatureExamples/Mods/Ragdoll2DHud/CPlusPlus/Data/NinjaSnowWar/Sight.png
