+++ FeatureExamples/Mods/Ragdoll2DHud/CPlusPlus/Data/hudfill1.png
