+++ FeatureExamples/Mods/Ragdoll2DHud/CPlusPlus/Data/hudfill2.png
