+++ FeatureExamples/Mods/Ragdoll2DHud/CPlusPlus/Data/hudmass.png
