+++ FeatureExamples/Mods/Ragdoll2DHud/CSharp/Resources/PostProcess/FrontPath.xml
