+++ FeatureExamples/Mods/Ragdoll2DHud/CSharp/Resources/Textures/hudfill2.png
