+++ FeatureExamples/Mods/Ragdoll2DHud/CSharp/Resources/Textures/hudfill6.png
