+++ FeatureExamples/Mods/Ragdoll2DHud/CSharp/Resources/Textures/hudmass.png
