+++ FeatureExamples/Mods/Ragdoll2DHud/CSharp/Resources/Textures/hudsize.png
