+++ FeatureExamples/Mods/Ragdoll2DHud/README.md
@@ -1,0 +1,6 @@
+
+#### Ragdoll Physics 2D HUD
+
+This example shows the technique of using a 2D viewport as a hud, having the same real estate as the 3D viewport.
+
+The non-clearing renderpath makes it all work, and then you just operate both scenes as separate but equal entities. I added the rotating coin to make sure animated sprites work. Using the "U" key, you can change the mass of the ball, "I" = Speed, "O" = Size, then you can see the appropriate HUD change in response to your selection. Then fire at will, or Jack, as it were. When Jack is hit, he gets the donuts blown out of him. Once all the Jacks are down, you can press "R" to restart the "level".
