+++ FeatureExamples/README.md
@@ -1,0 +1,12 @@
+
+#### Feature Examples
+
+This example showcases and provides example usage of specific Atomic technology features.
+
+The examples are minimal and strive to use minimal API and concise code.
+
+##### Feature Example Mods
+
+In the Mods subdirectory, are modifications of the feature example which show extended usage and subsystem integration.
+
+A mod can be applied, simply by copying the mod folder into your project's root folder.
